Revise fairseq2.nn modules
