+++ src/fairseq2/generation/_sampling/_generator.py
@@ -605,8 +605,8 @@ class _AbstractSamplingSequenceGeneratorOp(ABC):
 
         self._watch.reset()
 
-        self._counters.cache_size = self._state_bag.size_bytes()
-        self._counters.cache_capacity = self._state_bag.capacity_bytes()
+        self._counters.cache_size = self._state_bag.size()
+        self._counters.cache_capacity = self._state_bag.capacity()
 
         if self._compute_scores:
             # Sort the hypotheses by their scores before returning.
