+++ src/fairseq2/models/transformer_decoder/_model.py
@@ -16,7 +16,7 @@ from fairseq2.models.decoder import DecoderModel
 from fairseq2.models.sequence import SequenceModelOutput
 from fairseq2.models.transformer import TransformerFrontend
 from fairseq2.nn import IncrementalStateBag, Projection
-from fairseq2.nn.functional import CrossEntropy, cross_entropy
+from fairseq2.nn.ops import CrossEntropy, cross_entropy
 from fairseq2.nn.padding import PaddingMask
 from fairseq2.nn.transformer import TransformerDecoder
