+++ src/fairseq2/nn/transformer/_encoder_layer.py
@@ -19,6 +19,7 @@ from fairseq2.typing import DataType, Device
 
 # isort: split
 
+from fairseq2.nn import ResidualConnect, StandardResidualConnect
 from fairseq2.nn.transformer._attention_mask import AttentionMask
 from fairseq2.nn.transformer._ffn import FeedForwardNetwork
 from fairseq2.nn.transformer._layer_norm import (
@@ -27,7 +28,6 @@ from fairseq2.nn.transformer._layer_norm import (
 )
 from fairseq2.nn.transformer._multihead_attention import MultiheadAttention
 from fairseq2.nn.transformer._norm_order import TransformerNormOrder
-from fairseq2.nn.transformer._residual import ResidualConnect, StandardResidualConnect
 
 
 class TransformerEncoderLayer(Module, ABC):
